Consolidate role_required decorator
